compat+fix: _MAX_PATH alias, BITMAPFILEHEADER; fix (unsigned int)ptr casts
- windows.h: add #define _MAX_PATH MAX_PATH (MSVC underscore alias used by
  w3d_dep.cpp:283 for stack-allocated path buffer).
- windows.h: add BITMAPFILEHEADER struct (pragma pack(2), bfType/bfSize/
  bfReserved1/bfReserved2/bfOffBits) needed by ww3d.cpp BMP writer.
- surfaceclass.cpp: fix two pointer-to-integer casts that lose upper 32 bits
  on 64-bit: (unsigned int)lock_rect.pBits -> (uintptr_t)lock_rect.pBits.
  pBits is a void* D3D surface pointer; arithmetic must preserve full width.

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/surfaceclass.cpp
@@ -604,7 +604,7 @@ void SurfaceClass::FindBB(Vector2i *min,Vector2i*max)
 		for (x = min->I; x < max->I; x++) {
 
 			// HY - this is not endian safe
-			unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
+			unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
 			unsigned char myalpha=alpha[size-1];
 			myalpha=(myalpha>>(8-alphabits)) & mask;
 			if (myalpha) {
@@ -678,7 +678,7 @@ bool SurfaceClass::Is_Transparent_Column(unsigned int column)
 	for (y = 0; y < (int) sd.Height; y++)
 	{
 		// HY - this is not endian safe
-		unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+y*lock_rect.Pitch);		
+		unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+y*lock_rect.Pitch);
 		unsigned char myalpha=alpha[size-1];		
 		myalpha=(myalpha>>(8-alphabits)) & mask;		
 		if (myalpha) {
